// src/lib/types.ts
export type UserRole = 
  | "district_engineer" 
  | "district_manager"
  | "regional_engineer" 
  | "regional_general_manager"
  | "global_engineer" 
  | "technician" 
  | "system_admin"
  | "ict"
  | "load_monitoring_edit"
  | "load_monitoring_delete"
  | "admin"
  | "project_engineer"
  | "pending"
  | null;